+++ packages/common/src/constants/FeatureFlags.ts
@@ -39,6 +39,10 @@ export const FeatureFlags = {
     Model: {
       Dereference: 'ir.studio.model.dereference'
     },
+    Components: {
+      LegacyVolumetric: 'ir.studio.components.legacyVolumetric',
+      Volumetric: 'ir.studio.components.volumetric'
+    },
     Panel: {
       VisualScript: 'ir.editor.panel.visualScript',
       Portal: 'ir.editor.panel.portal',
