+++ packages/editor/src/functions/addMediaNode.ts
@@ -208,6 +208,7 @@ export async function addMediaNode(
     )
     return entityUUID
   } else if (url.includes('.uvol')) {
+    // TODO: detect whether to add LegacyVolumetricComponent or VolumetricComponent
     const { entityUUID } = EditorControlFunctions.createObjectFromSceneElement(
       [
         { name: VolumetricComponent.jsonID },
