+++ packages/ui/src/components/editor/properties/audio/positional/index.tsx
@@ -39,6 +39,7 @@ import NodeEditor from '@ir-engine/editor/src/panels/properties/common/NodeEdito
 import { SelectionState } from '@ir-engine/editor/src/services/SelectionServices'
 import { PositionalAudioComponent } from '@ir-engine/engine/src/audio/components/PositionalAudioComponent'
 import { DistanceModel, DistanceModelOptions } from '@ir-engine/engine/src/audio/constants/AudioConstants'
+import { LegacyVolumetricComponent } from '@ir-engine/engine/src/scene/components/LegacyVolumetricComponent'
 import { MediaComponent } from '@ir-engine/engine/src/scene/components/MediaComponent'
 import { VolumetricComponent } from '@ir-engine/engine/src/scene/components/VolumetricComponent'
 import Slider from '../../../../../primitives/tailwind/Slider'
@@ -55,7 +56,11 @@ export const PositionalAudioNodeEditor: EditorComponentType = (props) => {
   const audioComponent = useComponent(props.entity, PositionalAudioComponent)
 
   useEffect(() => {
-    if (!hasComponent(props.entity, MediaComponent) && !hasComponent(props.entity, VolumetricComponent)) {
+    if (
+      !hasComponent(props.entity, MediaComponent) &&
+      !hasComponent(props.entity, LegacyVolumetricComponent) &&
+      !hasComponent(props.entity, VolumetricComponent)
+    ) {
       const nodes = SelectionState.getSelectedEntities()
       EditorControlFunctions.addOrRemoveComponent(nodes, MediaComponent, true)
     }
